cleaned up HEAPS. fuelfurnace is done correctly

+++ Assets/Scripts/Item/FuelItem.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New Fuel Item", menuName = "Inventory/Fuel")]
 public class FuelItem : Item {
 
-    public float burnTicks = 160;   //8 seconds
+    public int burnTicks = 160;   //8 seconds
 }
